Move core.util stuff to hub
